Initial commit. Add README and BSD-2-Clause license.
This kicks off a new website project for the CHAOSS Project. As stated
in the README, this will likely become a Jekyll-powered blog and
documentation site with information and resources spanning multiple
CHAOSS Working Groups.

A BSD-2-Clause license was chosen to deliberately allow permissive
freedoms in how people rebuild the site configuration and source.
Generally, we want people to cite us as a reference if they use the work
in this repo somewhere else, and we are also OK with them citing the
CHAOSS Project as an example because we try to model best practices for
a wider community. In this context, the BSD-2-Clause best fits the
spirit and nature of the project at its conception.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>